Refactor: 추후 개발할 entity를 위해, 채팅방 레포지토리를 interface로 분리 + 인메모리 구현체 구현
