+++ src/main/java/com/f3/exercise_mate/chat/controller/ChatRoomController.java
@@ -1,7 +1,7 @@
 package com.f3.exercise_mate.chat.controller;
 
 import com.f3.exercise_mate.chat.dto.CreateRoomReqDTO;
-import com.f3.exercise_mate.chat.entity.ChatRoom;
+import com.f3.exercise_mate.chat.domain.ChatRoom;
 import com.f3.exercise_mate.chat.service.ChatRoomService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
