+++ src/main/java/com/f3/exercise_mate/chat/domain/ChatMessage.java
@@ -1,4 +1,4 @@
-package com.f3.exercise_mate.chat.entity;
+package com.f3.exercise_mate.chat.domain;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
